% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fn_tables.R
\name{All_FN_Tables}
\alias{All_FN_Tables}
\title{All_FN_Tables}
\usage{
All_FN_Tables()
}
\description{
A vector of tables to be imported from the DATA.ZIP that can be used by
other functions in the package.
}
