% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_spc_names.R
\name{append.spc.names}
\alias{append.spc.names}
\title{Append Species Names}
\usage{
append.spc.names(fntable)
}
\arguments{
\item{fntable}{a data table that has a column SPC indicating the species code (e.g. FN125)}
}
\value{
returns a merged dataframe
}
\description{
function that appends species names and codes (SpcName) to a working data frame.
}
\details{
append.spc.names
}
