% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_date.R
\name{fix_date}
\alias{fix_date}
\title{fix_date}
\usage{
fix_date(fndt0)
}
\arguments{
\item{fndt0}{A date field in FN2 files}
}
\value{
a formatted date field
}
\description{
A generic function to fix dates in FN2 files. Dates after 2000 are recorded as 19XX.
}
\examples{
fix_date(as.Date("1920-07-10")); fix_date(as.Date("1990-07-10")) 
}
