% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fl_hist.R
\name{fl.hist}
\alias{fl.hist}
\title{fl.hist}
\usage{
fl.hist(species, my125)
}
\arguments{
\item{species}{three digit species code (e.g. '081')}

\item{my125}{the location of the FN125 data (e.g. fndata$FN125)}
}
\description{
Fork length histogram
}
