% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fl_rwt.R
\name{fl.rwt}
\alias{fl.rwt}
\title{fl.rwt}
\usage{
fl.rwt(species, my125, logplot = FALSE)
}
\arguments{
\item{species}{three digit species code (e.g. '081')}

\item{my125}{the location of the FN125 data (e.g. fndata$FN125)}

\item{logplot}{optional argument to plot a log-log plot, default = FALSE}
}
\description{
fl.rwt
}
\details{
provides a quick diagnostic plot for looking for data errors
}
