% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fn_data1.R
\name{import.fn.data1}
\alias{import.fn.data1}
\title{Import a single FN project from a DATA.ZIP file}
\usage{
import.fn.data1(mydir, yr, program, mytables = all.tables)
}
\arguments{
\item{mydir}{should specify the local data warehouse ('c:/FN_Data')}

\item{yr}{is a field season year (e.g. 2013),}

\item{program}{is a character vector specifying which program to build a project code for ('GL1')}

\item{mytables}{is either a single FN2 table ('FN121') or a list of tables ('FN121', 'FN125').  If no value is supplied the default import routine is to import all supported tables.}
}
\value{
returns a list that contains each FN2 table as an item in the list
}
\description{
This function will generally not be called directly. It is mainly intended to be called from with import.fn.data
}
