% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_generic_fn_data.R
\name{import.generic.fn.data}
\alias{import.generic.fn.data}
\title{import.generic.fn.data}
\usage{
import.generic.fn.data(generic, mytables = all.tables)
}
\arguments{
\item{generic}{File path to FN DATA.ZIP folder}

\item{mytables}{Specify which FN data tables are to be imported.}
}
\value{
a list of FN2 tables
}
\description{
A generic import function will open and compile multiple FN projects.  
Project path build is omitted so it must be provided in file path.  
This function allows the import of projects other than 'GL1' and 'TW1' but doesn't not
format the outputs to the same extent.
}
