% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_generic_fn_data1.R
\name{import.generic.fn.data1}
\alias{import.generic.fn.data1}
\title{import.generic.fn.data1}
\usage{
import.generic.fn.data1(generic_datazip)
}
\arguments{
\item{generic}{File path to FN DATA.ZIP folder}
}
\description{
A generic import function similar to open.fn.data1.  Project path build is omitted so it must be provided in file path
}
