% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_creel_frtables.R
\name{import_creel_frtables}
\alias{import_creel_frtables}
\title{import creel frtables}
\usage{
import_creel_frtables(generic_datazip)
}
\arguments{
\item{generic_datazip}{a file path to a single FN2 DATA.ZIP creel project}
}
\value{
list a list with "FR712", "FR713", "FR714" and "FR715" tables
}
\description{
This function will import the FR report tables from a single FN2 creel project 
that has been analyzed in FN2. FN2 generates several FR* report/summary tables.
Tables will be returned as items within a list.
}
\examples{
\dontrun{
cr1_frtable <- import_creel_frtables{"SC97_WL1/DATA.ZIP"}
lapply(cr1_frtable, head)
}


}
\seealso{
[import_creel_series_frtables()] for combining multiple projects
}
