% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_creel_series.R
\name{import_creel_series}
\alias{import_creel_series}
\title{import_creel_series: Imports a series of DATA.ZIP files}
\usage{
import_creel_series(datazips)
}
\arguments{
\item{datazips}{}
}
\value{
list of FN2 Tables
}
\description{
import_creel_series: Imports a series of DATA.ZIP files
}
\examples{
\dontrun{
cr1 <- import_creel_series{"SC97_WL1/DATA.ZIP", "SC96_WL11/DATA.ZIP"}
}

}
