% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_creel_frtables.R
\name{import_creel_series_frtables}
\alias{import_creel_series_frtables}
\title{import creel series frtables}
\usage{
import_creel_series_frtables(datazips)
}
\arguments{
\item{datazips}{a vector of file paths to FN2 DATA.ZIP creel projects}
}
\value{
list a list with "FR712", "FR713", "FR714" and "FR715" tables
}
\description{
This function is a wrapper for `import_creel_frtables` that will import the FR 
report tables from multiple FN2 creel projects 
that have been analyzed in FN2. FN2 generates several FR* report/summary tables.
Tables will be returned as items within a list.
}
\examples{
\dontrun{
cr1_frtable <- import_creel_series_frtables("SC97_WL1/DATA.ZIP", "SC96_WL11/DATA.ZIP")
lapply(cr1_frtable, head)
}
}
\seealso{
[import_creel_frtables()] for combining multiple projects
}
