% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fn_creel1.R
\name{import_fn_creel}
\alias{import_fn_creel}
\title{import_fn_creel}
\usage{
import_fn_creel(generic_datazip)
}
\arguments{
\item{generic}{File path to FN DATA.ZIP folder}
}
\description{
A generic import function similar to open.fn.data1 but for creel files.  
Project path build is omitted so it must be provided in file path
}
