% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fn_index_net.R
\name{import_fn_index_net}
\alias{import_fn_index_net}
\title{import_fn_index_net}
\usage{
import_fn_index_net(generic_datazip)
}
\arguments{
\item{generic_datazip}{File path to FN DATA.ZIP folder}
}
\description{
A generic import function similar to open.fn.data1 but without interactivity.. 
Project path build is omitted so it must be provided in file path.
}
