% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_index_series.R
\name{import_index_series}
\alias{import_index_series}
\title{import_index_series: Imports a series of DATA.ZIP files}
\usage{
import_index_series(datazips)
}
\arguments{
\item{datazips}{}
}
\value{
list of FN2 Tables
}
\description{
import_index_series: Imports a series of DATA.ZIP files
}
\examples{
\dontrun{
gl1 <- import_index_series("IA21_GL1/DATA.ZIP", "IA22_GL1/DATA.ZIP")
}

}
