% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_raw_fn_data.R
\name{import_raw_fn_data}
\alias{import_raw_fn_data}
\title{Import Generic FN DATA.ZIP}
\usage{
import_raw_fn_data(generic_datazip)
}
\arguments{
\item{generic_datazip}{}
}
\value{
list of FN Tables
}
\description{
Provides a generic unzip function but with console prompts to add missing
FN011 fields.
}
\examples{
\dontrun{
fn <- import_raw_fn_data("~/FNData/NS1/IA02_NS1/DATA.ZIP")
}
}
