% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsf_fix_vartype.R
\name{lsf.vartypes}
\alias{lsf.vartypes}
\title{Assign LSF variable types}
\usage{
lsf.vartypes(lsftable)
}
\arguments{
\item{lsftable}{is a list containing each FN table as an item in the list}
}
\value{
returns a list that contains each FN2 table as an item in the list
}
\description{
Changes the default 'chr' class of each variable to the desired class type (e.g. 'int') for LSF data.  Generally this function will not be called directly as it is called from with 'import.fn.data'
}
