% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make.project}
\alias{make.project}
\title{Makes a FN2 project directory name}
\usage{
make.project(yr, program)
}
\arguments{
\item{yr}{is a field season year (e.g. 2013)}

\item{program}{is a character vector specifying a program ('GL1')}
}
\value{
returns a project directory code ('IA13_GL1')
}
\description{
Takes a year and a program type and forms a project directory code. Generally only used within other functions.
}
