% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns1_fix_vartype.R
\name{ns1.vartypes}
\alias{ns1.vartypes}
\title{Assign NS variable types}
\usage{
ns1.vartypes(ns1table)
}
\arguments{
\item{ns1table}{is a list containing each FN table as an item in the list}
}
\value{
returns a list that contains each FN table as an item in the list
}
\description{
Changes the default 'chr' class of each variable to the desired class type (e.g. 'int') for NS data.  Generally this function will not be called directly as it is called from within 'import.fn.data'
}
