% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_siloc.R
\name{SILOC2COORD}
\alias{SILOC2COORD}
\title{Convert XSILOC1 FN2 Field to LAT and LON fields}
\usage{
SILOC2COORD(fn121, fn121column)
}
\description{
XSILOC1 is a 12 character field that records LAT and LON in the format
DDMMddDDMMdd (LAT then LON). This function uses two helper functions to parse the field and add a
LAT and LON field to the FN121 table provided.
}
