% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcName.R
\docType{data}
\name{SpcName}
\alias{SpcName}
\title{Species Codes, Common Names and Species Labels}
\usage{
SpcName
}
\description{
FishNet data files rel on the use of codes to specify species.  This table provides a key between code and common name.
}
