% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_catch.R
\name{standardize.catch}
\alias{standardize.catch}
\title{standardize.catch}
\usage{
standardize.catch(NorW, fn121, fn123, netnights = FALSE, autofill = F)
}
\arguments{
\item{NorW}{specifies 'CATCNT' or 'CATWT'}

\item{fn121}{a valid FN121 table}

\item{fn123}{a valid FN123 table}

\item{netnights}{specifies whether catch totals should be standarizded to a 24 hour net set}

\item{autofill}{specifies whether standardize catch should run fill.CATCNT or fill.CATWT}
}
\description{
merges FN121 to FN123 and generates a catch table
}
