% Generated by roxygen2 (4.0.2): do not edit by hand
\name{import.generic.fn.data}
\alias{import.generic.fn.data}
\title{import.generic.fn.data}
\usage{
import.generic.fn.data(generic, mytables = all.tables)
}
\arguments{
\item{generic}{File path to FN DATA.ZIP folder}

\item{mytables}{Specify which FN data tables are to be imported.}
}
\description{
A generic import function similar to open.fn.data to open and compile multiple FN projects.  Project path build is omitted so it must be provided in file path.
}

