% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{SpcName}
\alias{SpcName}
\title{Species Codes, Common Names and Species Labels}
\usage{
SpcName
}
\description{
FishNet data files rel on the use of codes to specify species.  This table provides a key between code and common name.
}

