% Generated by roxygen2 (4.0.2): do not edit by hand
\name{standardize.catch}
\alias{standardize.catch}
\title{standardize.catch}
\usage{
standardize.catch(NorW, fn121, fn123, netnights = FALSE)
}
\arguments{
\item{NorW}{specifies 'CATCNT' or 'CATWT'}

\item{fn121}{a valid FN121 table}

\item{fn123}{a valid FN123 table}
}
\description{
merges FN121 to FN123 and generates a catch table
}

